+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Slideshow : fixed keyboard navigation, now the right arrow goes to the next slides and the left to the previous one.
+
 ## [2.1.3][] - 2021-11-19
 
 ### Added