+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Slideshow : fixed keyboard navigation, now the right arrow goes to the next slides and the left to the previous one.
 
+### Changed
+
+-   Soften medium and low dark buttons color.
+
 ## [2.1.3][] - 2021-11-19
 
 ### Added